5,203 AppleClang + 1,948 MSVC /W4 warning sites -> zero, enforced (-Wall
-Wextra -Werror GNU-family; /W4 /WX MSVC); -w deleted everywhere. ~35 real
bugs fixed (stack overflows, boot crashes, leaks, logic; all disclosed and
review-verified); sprintf conversion completed tree-wide; const/string_view
campaign; required ASan/UBSan CI jobs (6-job matrix); advisory .clang-tidy.
Goldens untouched across all 37 commits. Final review: READY TO MERGE 0C/0I.
Final CI: run 29214544984 six required jobs green.

+++ src/tests/boards_json_tests.cpp
@@ -3,6 +3,7 @@
 
 #include <gtest/gtest.h>
 
+#include <cstdint>
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
@@ -84,7 +85,7 @@ void append_record(std::string* bytes, int slot_num, int msg_num, int level, int
     // Garbage, deliberately non-null and never a valid pointer this process
     // owns -- proves the decoder reads past this field and discards it
     // rather than dereferencing it.
-    record.heading = reinterpret_cast<char*>(0xdeadbeef);
+    record.heading = reinterpret_cast<char*>(static_cast<std::uintptr_t>(0xdeadbeef));
     record.level = level;
     record.post_time = post_time;
     record.heading_len = static_cast<int>(heading.size()) + 1;
@@ -244,7 +245,7 @@ TEST(BoardsJson, RejectsZeroHeadingLenAsCorrupt)
     board_msginfo record {};
     record.slot_num = 1;
     record.msg_num = 1;
-    record.heading = reinterpret_cast<char*>(0xdeadbeef);
+    record.heading = reinterpret_cast<char*>(static_cast<std::uintptr_t>(0xdeadbeef));
     record.level = 1;
     record.post_time = 1;
     record.heading_len = 0;
